# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EngTv IPTV API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID with its channels
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category with channels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryWithChannels"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Public: Channels ─────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login — returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin: Settings ──────────────────────────────────────────────────────
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [categories, admin]
      summary: List all categories including hidden (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories, admin]
      summary: Create a category (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/reorder:
    patch:
      operationId: reorderCategories
      tags: [categories, admin]
      summary: Reorder categories (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories, admin]
      summary: Update a category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: [categories, admin]
      summary: Delete a category (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin: Channels ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── App Settings ──────────────────────────────────────────────────────────
...
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
    AdminLoginResult:
      type: object
      required: [token]
      properties:
        token:
          type: string

    # ── Category ───────────────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name, sort_order, is_visible]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        sort_order:
          type: integer
        is_visible:
          type: boolean
        created_at:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: ["string", "null"]
        sort_order:
          type: ["integer", "null"]
        is_visible:
          type: boolean
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: ["string", "null"]
        sort_order:
          type: ["integer", "null"]
        is_visible:
          type: boolean
    CategoryWithChannels:
      type: object
      required: [id, name, sort_order, is_visible, channels]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        sort_order:
          type: integer
        is_visible:
          type: boolean
        created_at:
          type: string
        channels:
          type: array
          items:
            $ref: "#/components/schemas/Channel"

    # ── Channel ────────────────────────────────────────────────────────────────
    Channel:
      type: object
      required: [id, name, stream_url, is_active, sort_order]
      properties:
        id:
          type: integer
        name:
          type: string
        stream_url:
          type: string
        logo_url:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        source_id:
          type: ["integer", "null"]
        external_id:
          type: ["string", "null"]
        is_active:
          type: boolean
        sort_order:
          type: integer
        language:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        created_at:
          type: string
        last_checked_at:
          type: ["string", "null"]
        is_healthy:
          type: ["boolean", "null"]
        health_error:
          type: ["string", "null"]
...
    ReorderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReorderItem"
    ReorderItem:
      type: object
      required: [id, sort_order]
      properties:
        id:
          type: integer
        sort_order:
          type: integer

    # ── Source ─────────────────────────────────────────────────────────────────
...